refactor: Edit dialog variable name
