+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/UiState.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.home.presentation
+package com.teamoffroad.feature.home.presentation.component
 
 sealed interface UiState<out T> {
