+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/model/ErrorMessageModel.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.home.presentation
+package com.teamoffroad.feature.home.presentation.model
 
 data class ErrorMessageModel(
     val message: String,
